Refactor for looser coupling between main library and parsers
